% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jervis_bardy.R
\name{CalcSpearmanMatrix}
\alias{CalcSpearmanMatrix}
\title{CalcSpearmanMatrix() Function}
\usage{
CalcSpearmanMatrix(otuMatrix, intensityVector)
}
\arguments{
\item{otuMatrix}{otu_table(phyloseq_object) works fine}

\item{intensityVector}{list of concentrations in the same order, aka: sample_data(phyloseq_object)$Concentration}
}
\description{
Function to calculate spearman correlation of OTU abundance to DNA concentration
Thanks to David Waite for the original verison of this code. :)
}
\details{
Returns a list of 3 matrices:
 * retMatrix -- full matrix of results (Rho, P.value, FDR)
 * negSigMatrix -- subset to stat sig (p < 0.05) and negatively correlated OTUs
 * posSigMatrix -- subset to stat sig (p < 0.05) and positively correlated OTUs
}
\examples{
JB_corr <- CalcSpearmanMatrix(phyloseq::otu_table(ps),phyloseq::sample_data(ps)$DNA_Concentration)
}
\keyword{contamination}
\keyword{correlation,}
\keyword{jervis-bardy,}
\keyword{spearman,}
