% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomic_functions.R
\name{CreateRelabund}
\alias{CreateRelabund}
\title{CreateRelabund Function}
\usage{
CreateRelabund(physeq, filt = FALSE)
}
\arguments{
\item{physeq}{phyloseq object to convert}

\item{filt}{minimum mean value to filter to (default is FALSE (no filtering))}
}
\description{
Function to generate relative-abundance table versions of a phyloseq object
Returns a list of phyloseq objects:
ret$orig = the original phyloseq object submitted
ret$r = the relative-abundance version
ret$fr = the filtered relative abundance version (if specified)
}
\examples{
ps <- CreateRelabund(physeq_object)
ps$orig; ps$r
ps <- CreateRelabund(physeq_object, filt=1e-5)
ps$orig; ps$r; ps$fr
}
\keyword{abundance,}
\keyword{phyloseq}
\keyword{relative}
