% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tables.R
\name{CreateTables}
\alias{CreateTables}
\title{CreateTables Function}
\usage{
CreateTables(physeq, metadata, cnames, key = "SampleID", otu_map = NULL)
}
\arguments{
\item{physeq}{phyloseq object to convert}

\item{metadata}{metadata table to merge}

\item{cnames}{char vector of metadata column names to include in output}

\item{key}{name of key variable (default = "SampleID")}

\item{otu_map}{optional: df["Seq","SeqID"] pairing sequence ID to its sequence}
}
\description{
Function to convert phyloseq objects into a set of
human-readable OTU tables analagous to QIIME output
will also append sequences to master table if requested
}
\examples{
cnames = c("SampleID","SampleType","TimePoint","Concentration","Batch","Description")
ret <- WriteTables(ps$orig, Metadata, cnames, key="SampleID")
write.csv(ret$master, "master_table.csv")
write.csv(ret$sum, "summary_table.csv")
write.csv(ret$otu, "otu_table.csv")
write.csv(ret$tax, "tax_table.csv")
cat(sapply(ret$fasta, toString), file="otu_mapping.fasta", sep="")
}
\keyword{dada2,}
\keyword{otu_table}
\keyword{phyloseq,}
\keyword{qiime,}
