% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tables.R
\name{CreateTablesSeqIDs}
\alias{CreateTablesSeqIDs}
\title{CreateTablesSeqIDs Function}
\usage{
CreateTablesSeqIDs(physeq, metadata, cnames, key = "SampleID")
}
\arguments{
\item{physeq}{phyloseq object to convert}

\item{metadata}{metadata table to merge}

\item{cnames}{char vector of metadata column names to include in output}

\item{key}{name of key variable (default = "SampleID")}
}
\description{
Function to convert phyloseq objects of DADA2 output
into a set of human-readable OTU tables analagous to QIIME output,
including creating sequence ID's & a fasta mapping for the taxa found
}
\examples{
cnames = c("SampleID","SampleType","TimePoint","Description")
ret <- WriteTablesSeqIDs(ps$orig, sample_data(ps$orig), cnames, key="SampleID")
write.csv(ret$master, "master_table.csv")
write.csv(ret$sum, "summary_table.csv")
write.csv(ret$otu, "otu_table.csv")
write.csv(ret$tax, "tax_table.csv")
cat(sapply(ret$fasta, toString), file="otu_mapping.fasta", sep="")
}
\keyword{dada2,}
\keyword{otu_table}
\keyword{phyloseq,}
\keyword{qiime,}
