% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{class}
\name{evalObj-class}
\alias{evalObj-class}
\alias{evalObj}
\title{A custom class for storing the chromatographic peak data required by the peak metric functions for each group of samples.}
\description{
A custom class for storing the chromatographic peak data required by the peak metric functions for each group of samples.
}
\section{Slots}{

\describe{
\item{\code{eicPts}}{A list of 2D matrices containing the retention time and intensity values of each chromatographic peak}

\item{\code{eicPeakData}}{A list of vectors for each sample in the group containing characteristic information about each chromatographic peak}

\item{\code{eicNos}}{A numeric vector of the EIC numbers identifying each feature group}
}}

