% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_peakData}
\alias{ex_peakData}
\title{Example peakData - value input to calculate... functions (except calculateElutionShift and calculateRetentionTimeConsistency)}
\format{
A list containing the following entries: \code{mz}, \code{mzmin}, \code{mzmax}, \code{rt}, \code{rtmin}, \code{rtmax}, \code{into},
\code{intb}, \code{maxo}, \code{sn}, \code{sample}, and \code{is_filled}.
}
\usage{
ex_peakData
}
\description{
An example of the input for the peakData argument for calculate... functions. It represents data from one sample for the peak of
interest.
}
\keyword{datasets}
