% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_pts}
\alias{ex_pts}
\title{Example pts - value input to caculate... functions (except calculateElutionShift and calculateRetentionTimeConsistency)}
\format{
A two-column matrix where the first column represents \code{rt} and the second column represents
\code{intensity}.
}
\usage{
ex_pts
}
\description{
An example of the input for the pts argument for calcualte... functions. It represents rt and intensity data from one sample for
peak of interest.
}
\keyword{datasets}
