% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_ptsList}
\alias{ex_ptsList}
\title{Example ptsList - value input to calculteElutionShift and calculateRetentionTimeConsistency}
\format{
A list of two-column matrices (one matrix per sample) where the first column represents \code{rt} and the second column
 represents \code{intensity}.
}
\usage{
ex_ptsList
}
\description{
An example of the input for the ptsList argument for calculteElutionShift and calculateRetentionTimeConsistency. Each entry in the
list is a two-column matrix consisting of rt and intensity for a sample for the peak of interest.
}
\keyword{datasets}
