% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomic_functions.R
\name{ExtendPlastidTax}
\alias{ExtendPlastidTax}
\title{ExtendPlastidTax Function}
\usage{
ExtendPlastidTax(ps)
}
\arguments{
\item{physeq}{phyloseq object to correct}
}
\description{
The dada2 silva_v128 taxonomy database only classifies mitochondria to
Family and chloroplasts to Class level. This is problematic when these taxa 
have (deliberately) not been removed from the dataset, are present in 
significant amounts, and one wants to look at distributions at lower
taxonomy levels.
}
\details{
This function takes a phyloseq object with this problem, and fills in 
the taxonomy table for these plastids down the the lowest level
(either Genus or Species)
}
\examples{
ps <- ExtendPlastids(ps)
#to check that it worked:
df <- as.data.frame(tax_table(ps))
filter( df[which(df$Family == "Mitochondria"), ] )[1:3, ]
}
\keyword{dada2,}
\keyword{phyloseq,}
\keyword{taxonomy}
