% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEvalObj.R
\name{getEvalObj}
\alias{getEvalObj}
\title{Extract peak data object}
\usage{
getEvalObj(xs, fill)
}
\arguments{
\item{xs}{An xcmsEIC object returned by the getEIC() function from the XCMS package}

\item{fill}{An xcmsSet object with filled in peak groups}
}
\value{
An object of class evalObj
}
\description{
This function extracts, formats, and combines the chromatographic peak data from the objects returned by the getEIC() and fillPeaks()
functions from the XCMS package.
}
\examples{
# call getEvalObj on test data
# \donttest{eicEval_test <- getEvalObj(xs = xs_test, fill = fill_test)}

}
