% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqcFilter.R
\name{lqcFiltering}
\alias{lqcFiltering}
\title{LQC Filteirng}
\usage{
lqcFiltering(peakTable, eicColumn, lqcColumns, rsdThreshold = 0.3)
}
\arguments{
\item{peakTable}{peak table generated by xcms group object}

\item{eicColumn}{name of the EIC column}

\item{lqcColumns}{names of the LQC columns}

\item{rsdThreshold}{RSD % threshold for filtering; default 0.3}
}
\value{
peakTable with filtered EICs
}
\description{
Filters out EICs with LQC RSD % above a user provided threshold.
}
\examples{
\donttest{
   lqc_filtered_table <- lqcFiltering(peakTable = group_table,
}

 @export
}
