% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pqm_development}
\alias{pqm_development}
\title{Example Peak Quality Metrics Data Frame for Development Dataset.}
\format{
A data frame with 13 variables (EIC Number, the 11 peak quality metrics, and Class Labels): \code{EICNo},
\code{ApexBoundaryRatio_mean}, \code{ElutionShift_mean}, \code{FWHM2Base_mean}, \code{Jaggedness_mean}, \code{Modelaity_mean},
\code{RetentionTimeCorrelation_mean}, \code{Symmetry_mean}, \code{GaussianSimilarity_mean}, \code{Sharpness_mean},
\code{TPASR_mean}, \code{ZigZag_mean}, and \code{Class}.
}
\usage{
pqm_development
}
\description{
Data frame with peaks quality metrics and labels for all of the 500 EICs in the example development dataset.
}
\keyword{datasets}
