% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pqMetrics_test}
\alias{pqMetrics_test}
\title{Example Peak Quality Metrics Data Frame for Test Dataset.}
\format{
A data frame with 14 variables (EIC Number, the 12 peak quality metrics, and Class Labels): \code{EICNo},
\code{ApexBoundaryRatio_mean}, \code{ElutionShift_mean}, \code{FWHM2Base_mean}, \code{Jaggedness_mean}, \code{Modelaity_mean},
\code{RetentionTimeCorrelation_mean}, \code{Symmetry_mean}, \code{GaussianSimilarity_mean}, \code{PeakSignificance_mean},
\code{Sharpness_mean}, \code{TPASR_mean}, \code{ZigZag_mean}, and \code{Class}.
}
\usage{
pqMetrics_test
}
\description{
Data frame with peaks quality metrics and labels for all of the 500 EICs in the example test dataset.
}
\keyword{datasets}
