% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomic_functions.R
\name{ReplaceSeqIDs}
\alias{ReplaceSeqIDs}
\title{ReplaceSeqIDs Function}
\usage{
ReplaceSeqIDs(physeq, fastafile)
}
\arguments{
\item{physeq}{phyloseq object to convert}

\item{name}{of fasta file to write to}
}
\description{
Replace dada2 sequence strings with seqence ID values
and generate a .fasta file to match these to their sequences.
Use to make working with dada2 output a bit easier
}
\examples{
ret <- ReplaceSeqIDs(ps, "seqids.fa")
ps2 <- ret$physeq
otu_map <- ret$otu_map
system("head -4 seqids.fa", intern=TRUE) #peek at the fasta file
}
\keyword{dada2,}
\keyword{otu_table}
\keyword{phyloseq,}
