% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomic_functions.R
\name{SubsetSamples}
\alias{SubsetSamples}
\title{SubsetSamples Function}
\usage{
SubsetSamples(physeq, samplelist)
}
\arguments{
\item{physeq}{phyloseq object to subset}

\item{samplelist}{character vector of samples to keep}
}
\description{
Function to subset a phyloseq object to a specific subset of samples,
and remove taxa not found in those samples from the phyloseq object.
Especially useful for separating out blank samples
}
\examples{
ps_blanks <- SubsetSamples(ps, blank_samples)
}
\keyword{blanks}
\keyword{filter,}
\keyword{subsettaxa,}
